WIP: [Nonlinear] begin experiments with NonlinearExpr
This PR is an experiment with different representations of a nonlinear
expression type for JuMP.

As a naive first pass, we replicate the structure of basic Julia Expr
objects. The motivation for this is that we already create these in the
macros, and it seems to work okay, so why not make it official! Once I
get benchmarks setup, etc, we may need to try something else, but this is
a good reference implementation.